+++ R/blend_predictions.R
@@ -57,7 +57,7 @@
 #' 
 #' @template note_example_data
 #' 
-#' @examples 
+#' @examplesIf rlang::is_installed("kernlab")
 #' \donttest{
 #' # see the "Example Data" section above for
 #' # clarification on the objects used in these examples!
@@ -140,6 +140,7 @@ blend_predictions <- function(data_stack,
   }
   check_inherits(control, "control_grid")
   check_inherits(times, "numeric")
+  check_empty_ellipses(...)
   
   outcome <- attr(data_stack, "outcome")