+++ R/fit_members.R
@@ -20,11 +20,7 @@
 #' 
 #' @template note_example_data
 #' 
-#' @examples 
-#' # feel free to ignore the following lines—they allow {stacks} to supply 
-#' # examples without requiring the model-supplying package to be installed.
-#' if (requireNamespace("ranger", quietly = TRUE)) {
-#'   if (requireNamespace("kernlab", quietly = TRUE)) {
+#' @examplesIf rlang::is_installed("ranger") & rlang::is_installed("kernlab")
 #' 
 #' \donttest{
 #' 
@@ -66,14 +62,15 @@
 #'   fit_members()
 #'   
 #' log_st
-#'     }
-#'   }
+#' 
 #' }
 #' 
 #' @family core verbs
 #' @export
 fit_members <- function(model_stack, ...) {
   check_model_stack(model_stack)
+  check_for_required_packages(model_stack)
+  check_empty_ellipses(...)
   
   dat <- model_stack[["train"]]
   
@@ -225,4 +222,60 @@ check_model_stack <- function(model_stack) {
   } else {
     check_inherits(model_stack, "model_stack")
   }
+}
+
+# given a model stack, find the packages required to fit members and predict 
+# on new values, and error if any of them are not loaded
+check_for_required_packages <- function(x) {
+  # for dispatch to required_pkgs.workflow when model
+  # is loaded in a fresh environment
+  suppressPackageStartupMessages(requireNamespace("workflows"))
+  
+  pkgs <-
+    purrr::map(
+      x$model_defs,
+      parsnip::required_pkgs
+    ) %>%
+    unlist() %>%
+    unique()
+
+  installed <- purrr::map_lgl(
+    pkgs,
+    is_installed_
+  )
+  
+  if (any(!installed)) {
+    error_needs_install(pkgs, installed)
+  }
+
+  purrr::map(
+    pkgs,
+    ~suppressPackageStartupMessages(requireNamespace(.x, quietly = TRUE))
+  )
+  
+  invisible(TRUE)
+}
+
+# takes in a vector of package names and a logical vector giving
+# whether or not each is installed
+error_needs_install <- function(pkgs, installed) {
+  plural <- sum(!installed) != 1
+  
+  last_sep <- if (sum(!installed) == 2) {"` and `"} else {"`, and `"}
+  
+  need_install <- paste0(
+    "`",
+    glue::glue_collapse(pkgs[!installed], sep = "`, `", last = last_sep),
+    "`"
+  )
+  
+  glue_stop(
+    "The following package{if (plural) 's' else ''} ",
+    "need{if (plural) '' else 's'} to be installed before ",
+    "fitting members: {need_install}"
+  )
+}
+
+is_installed_ <- function(pkg) {
+  rlang::is_installed(pkg)
 }